requirements.txt
python-dotenv~=1.1.1
pymongo~=4.13.2
motor~=3.7.0
RapidFuzz~=3.13.0
pytz~=2025.2
Unidecode~=1.4.0
tabulate~=0.9.0
uvicorn~=0.34.3
uvicorn
itsdangerous
starlette~=0.40.0
pendulum~=2.1.0
aiohttp~=3.11.18
fuzzywuzzy~=0.18.0
discord.py~=2.5.2
pillow~=11.3.0